chat-with-pdfs/requirements.txt
langchain
openai
PyPDF2
faiss-cpu
huggingface-hub
python-dotenv
sentence-transformers
langchain-community